Fixed a windows incompatibility
Thanks to tumblr.com/pollstuck for bringing this to my attention. Apparently, some Windows installations of Python use CP1252 instead of UTF-8 as the default encoding.  This patch ensures that UTF-8 is used to read the mspa.json file.  This bug is already patched in the beta branch, but that is not yet complete, so this slipped by!